Extract methods for validating titles
